pyproject.toml
[tool.poetry]
name = "python-bot"
version = "0.1.0"
description = ""
authors = ["hungdhv97 <hungdhv97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.2"
requests = "^2.30.0"
bs4 = "^0.0.1"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"
pycryptodome = "^3.17"
keyring = "^23.13.1"
pip = "^23.1.2"
cryptography = "^40.0.2"
secretstorage = "^3.3.3"
setuptools = "^67.7.2"
wheel = "^0.40.0"
pywin32 = { version = "^306", platform = "win32" }
pydantic = "^1.10.7"
pytest = "^7.3.1"
pynput = "^1.7.6"
pyjwt = "^2.7.0"
pyasn1 = "^0.5.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
types-requests = "^2.30.0.0"
types-urllib3 = "^1.26.25.12"
types-pynput = "^1.7.5.6"
types-beautifulsoup4 = "^4.12.0.5"
types-pillow = "^9.5.0.4"
pyinstaller = { version = "^5.10.1", python = "3.10" }
mypy = "^1.3.0"
mypy-extensions = "^1.0.0"
invoke = "^2.1.1"
pylint = "^2.17.4"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.3"

[tool.flake8]
exclude = ["venv"]
ignore = ['E231', 'E241', "E501", "E203"]
max-line-length = "88"

[tool.pylint]
disable = ["C0114", "C0115", "C0116", "C0415", "R0903"]
ignore-patterns = ["venv/*"]
recursive = "true"
ignored-modules="win32crypt"
max-locals=25

[tool.black]
line-length = "88"
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = "88"
multi_line_output = "3"
include_trailing_comma = "true"
force_grid_wrap = "0"
combine_as_imports = "true"

[tool.mypy]
exclude = ['venv/']
python_executable = "venv/bin/python"
explicit_package_bases = "true"
check_untyped_defs = "true"
python_version = "3.10"
warn_return_any = "true"
warn_unused_configs = "true"
strict_optional = "true"
disallow_untyped_defs = "true"
disallow_untyped_calls = "true"
no_implicit_optional = "true"
strict_equality = "true"
warn_no_return = "true"
disallow_any_generics = "true"
ignore_missing_imports = "true"
ignore_missing_imports_per_module = "true"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
